+++ examples/app/handlers/group1/group1.event.js
@@ -1,0 +1,5 @@
+module.exports = function Handler(log) {
+  this.handle = async function(ctx) {
+    ctx.success()
+  }
+}